feat(glog): 初始化时只存储配置
